Changed how CQI are calculated - Fixes #94 (#118)
* Changed how CQI are calculated - Fixes #94

Now CQI is calculated from:
1. cell tx power in dB plus
2. distance related attenuation in dB plus
3. angle related attenuation in dB

some gross simplifications are made, but it should work better with MLB app

Also used 'double' datatype in proto file to give 'float64' in go code
This spares a lot of conversion and gives better resolution to decimals

* Corrected the order of strongest

* Added more tests

+++ pkg/config/towerconfig.go
@@ -29,15 +29,15 @@ type Sector struct {
 	Azimuth     uint16     `yaml:"azimuth"`
 	Arc         uint16     `yaml:"arc"`
 	MaxUEs      uint16     `yaml:"maxues"`
-	InitPowerDb float32    `yaml:"initpowerdb"`
+	InitPowerDb float64    `yaml:"initpowerdb"`
 }
 
 // TowersLayout an individual tower with sectors
 type TowersLayout struct {
 	TowerID   string       `yaml:"towerid"`
 	PlmnID    types.PlmnID `yaml:"plmnid"`
-	Latitude  float32      `yaml:"latitude"`
-	Longitude float32      `yaml:"longitude"`
+	Latitude  float64      `yaml:"latitude"`
+	Longitude float64      `yaml:"longitude"`
 	Sectors   []Sector
 }
